Early prototype with WolfSSL and ALPN implementation
This commit has first code that migrated to WolfSSL and ALPN implemented. Rough but it should work enough for testing. It has mappings to WolfSSL user settings so you only need to compile WolfSSL with ALPN Support, set the library and include paths and then this code should compile. Ref:#1