Cleaned up code and made it got it's last shape with WolfSSL and ALPN
Removed leftovers from OpenSSL and replaced temporary debug code. Code is now as stable as old code which was powered with OpenSSL. Code structure is also changed; moved 'parseHeader' and 'serverHeaders' to 'AlyssaHTTP' class like it should be because all that will be rewritten from scratch on HTTP/2, And as of now the 'clientConnection' is unified so which means theres no separate 'clientConnection_SSL' anymore, single function will deal with both and still can be compiled with/without SSL support without any change in code. Ref: #1