Initial HPack code and "functional" HTTP/2 code
This commit has the initial HPack code. It is semi functional, meanwhile
it can parse most examples on RFC 7541, it's still not enough for
parsing actual HPack data, and dynamic table is partly implemented. This
commit also has first functional HTTP/2 code, which has accepting
connections and responding with a dummy page. Ref: #1