I ain't dead!
Had to take a break because of school exams and some other reasons. With this commit now the custom actions are collected in a single .alyssa file rather than separate files for every nodes. Recursive custom actions are now implemented (in theory) and is completely rewritten. This commit also includes some other improvements for other things that going to be changed.