3.0 Prerelease 1.2
This release makes drastical changes on how HTTP/2 works, now instead of
saving every stream on expanding deque, it now saves only 8 *active* streams
on a fixed array. Which solves wasted ram and also potential vulnerability
for RAM exhausting. This array may be made dynamic (dynamic as configurable to a value, not in an expanding on runtime fashion).
Also as a resuld, number of concurrent streams is limited to 8 or in future whatever that value is set to by user.
There is also some few improvements for HTTP/2.

Asides that, an error in allocation that is allocating stupid amount of memory is fixed.